Исправлено удаление контактов

+++ app/Http/Controllers/MemberContactController.php
@@ -111,7 +111,9 @@ class MemberContactController extends Controller
     public function destroy(Contact $contact)
     {
         $this->authorize('delete', $contact);
-        Storage::delete($contact->avatar);
+        if ($contact->avatar != 'avatar/no-image.jpg') {
+            Storage::delete($contact->avatar);
+        }
         $contact->delete();
         return redirect()->route('member.contacts.index');
     }
